refactor: move tests and types under src for easier ts build
