+++ src/extensions/fileWriter.ts
@@ -1,11 +1,11 @@
 import { dirname, resolve } from 'path';
 import { writeFileSync, mkdirSync } from 'fs';
 import { Canvas } from 'canvas';
-import { fileWriterOptions } from '../types';
+import { fileWriterOptions } from '../@types';
 
 export default (options?: fileWriterOptions) => (canvas: Canvas) => {
   const fileName =
-    options?.fileName ||
+    options?.fileName ??
     `${new Date().toISOString().replace(/[\W.]/g, '')}.png`;
 
   mkdirSync(resolve(dirname(fileName)), { recursive: true });
