+++ src/tests/helpers/extractColors.ts
