+++ src/tests/helpers/readImageData.ts
