+++ src/tests/textToImage.test.ts
@@ -7,15 +7,14 @@ import {
   countWhitePixels,
 } from './helpers/readImageData';
 import longInput from './helpers/longInput';
-import { generate, generateSync } from '../src';
+import { generate, generateSync } from '..';
 import takeSnapshot from './helpers/takeSnapshot';
 import { Canvas } from 'canvas';
-import { ComputedOptions } from '../src/types';
+import { ComputedOptions } from '../@types';
 
 describe('the text-to-image generator', () => {
-  it('should return a promise', (done) => {
+  it('should return a promise', () => {
     expect(generate('Hello world')).toBeInstanceOf(Promise);
-    done();
   });
 
   it('should have a sync version', () => {
